requirements.txt
openai
python-dotenv
spacy
pdfplumber
python-docx
